requirements/requirements.txt
# try to install a minimum set of packages only
# https://pypi.org/project/PyYAML/
pyyaml>=6.0.2
# https://pypi.org/project/python-dateutil/#history
python-dateutil >= 2.9.0
# use beautiful soup to parse markup documents 
beautifulsoup4 >= 4.14.2
# https://pypi.org/project/tzdata/
tzdata >= 2025.2
# https://pypi.org/project/lxml/
lxml >=6.0.2
# https://pypi.org/project/requests/
requests >= 2.32.5
# https://pypi.org/project/pywin32/
# https://superuser.com/questions/609447/how-to-install-the-win32com-python-library
# right now only used in legacy module
pywin32 >= 311
